# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "用户中心服务 API 文档",
        "title": "User Hub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/api/v1/user-hub/profile": {
            "get": {
                "description": "获取当前认证用户的核心账户信息（如角色、状态）和详细个人资料（如昵称、头像）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资料管理 (Profile Management)"
                ],
                "summary": "获取我的账户详情 (核心信息 + 资料)",
                "responses": {
                    "200": {
                        "description": "获取账户详情成功",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIMyAccountDetailResponse"
                        }
                    },
                    "401": {
                        "description": "未授权或认证失败",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据不一致或数据库查询失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            },
            "put": {
                "description": "当前认证用户更新自己的个人资料信息（如昵称、性别、地区等）。头像更新请使用专门的头像上传接口。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资料管理 (Profile Management)"
                ],
                "summary": "更新我的用户资料",
                "parameters": [
                    {
                        "description": "包含待更新字段的资料信息（不含头像URL）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateProfileDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "资料更新成功，返回更新后的资料信息",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIProfileVOResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如JSON格式错误)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "401": {
                        "description": "未授权或认证失败",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库操作失败或用户资料不存在)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        },
        "/api/v1/user-hub/profile/avatar": {
            "post": {
                "description": "当前认证用户上传自己的头像文件。成功后返回新的头像URL。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资料管理 (Profile Management)"
                ],
                "summary": "上传我的头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "头像文件 (multipart/form-data key: 'avatar')",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "头像上传成功，返回包含新头像URL的map",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-map_string_string"
                        }
                    },
                    "400": {
                        "description": "请求无效 (如文件过大、类型不支持、未提供文件)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "401": {
                        "description": "未授权或认证失败",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如文件上传到COS失败、数据库更新失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user-hub/users/{userID}/profile": {
            "get": {
                "description": "(管理员权限) 根据提供的用户ID，获取该用户的详细个人资料信息（昵称、头像等）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理 (User Management)"
                ],
                "summary": "获取指定用户资料 (管理员)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询的用户ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户资料成功",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIProfileVOResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如用户ID为空)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "403": {
                        "description": "权限不足 (非管理员操作)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "404": {
                        "description": "指定用户的资料不存在",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库查询失败)",
                        "schema": {
                            "$ref": "#/definitions/docs.SwaggerAPIErrorResponseString"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.SwaggerAPIErrorResponseString": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "type": "string"
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "docs.SwaggerAPIMyAccountDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.MyAccountDetailVO"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "docs.SwaggerAPIProfileVOResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.ProfileVO"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.UpdateProfileDTO": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "城市 (可选更新)",
                    "type": "string",
                    "example": "深圳"
                },
                "gender": {
                    "description": "性别（0=未知, 1=男, 2=女）(可选更新)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.Gender"
                        }
                    ],
                    "example": 1
                },
                "nickname": {
                    "description": "昵称 (可选更新)",
                    "type": "string",
                    "example": "小明"
                },
                "province": {
                    "description": "省份 (可选更新)",
                    "type": "string",
                    "example": "广东"
                }
            }
        },
...
        "enums.Gender": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "Female": "女性",
                "Male": "男性",
                "Unknown": "未知"
            },
            "x-enum-varnames": [
                "Unknown",
                "Male",
                "Female"
            ]
        },
...
        "enums.UserRole": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "RoleAdmin": "0 - 管理员，具有最高权限",
                "RoleGuest": "2 - 访客，限制性访问权限",
                "RoleUser": "1 - 普通用户，标准用户角色"
            },
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser",
                "RoleGuest"
            ]
        },
        "enums.UserStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-comments": {
                "StatusActive": "活跃，用户可以正常操作",
                "StatusBlacklisted": "拉黑，用户被禁止访问"
            },
            "x-enum-varnames": [
                "StatusActive",
                "StatusBlacklisted"
            ]
        },
        "map_string_string": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "response.APIResponse-map_string_string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/map_string_string"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "vo.MyAccountDetailVO": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "city": {
                    "type": "string",
                    "example": "深圳"
                },
                "created_at": {
                    "description": "可以是 User 的创建时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "gender": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.Gender"
                        }
                    ],
                    "example": 1
                },
                "nickname": {
                    "description": "来自 UserProfile 实体",
                    "type": "string",
                    "example": "小明"
                },
                "province": {
                    "type": "string",
                    "example": "广东"
                },
                "status": {
                    "description": "来自 User 实体",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.UserStatus"
                        }
                    ],
                    "example": 0
                },
                "updated_at": {
                    "description": "可以是 User 或 Profile 中较新的更新时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "user_role": {
                    "description": "来自 User 实体",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.UserRole"
                        }
                    ],
                    "example": 1
                }
            }
        },
        "vo.ProfileVO": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "头像 URL",
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "city": {
                    "description": "城市",
                    "type": "string",
                    "example": "深圳"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "gender": {
                    "description": "性别（0=未知, 1=男, 2=女）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.Gender"
                        }
                    ],
                    "example": 1
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string",
                    "example": "小明"
                },
                "province": {
                    "description": "省份",
                    "type": "string",
                    "example": "广东"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
    }
}